# backend/swagger.yaml
openapi: 3.0.0
info: 
  title: Order Creation API Documentation
  description: Outlines all the routes in our order creation API
  version: 1.0.0
paths:
...
  /v1/user/register:
    post:
      summary: Register a new user
      description: Registers a new user and returns a session token if successful
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "Password123"
                nameFirst:
                  type: string
                  example: "John"
                nameLast:
                  type: string
                  example: "Smith"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "1t2nb34"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
  /v1/user/login:
    post:
      summary: Login a user
      description: Logs in a user and returns a session token if successful
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "Password123"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "1t2nb34"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
  /v1/user/logout:
    post:
      summary: Logout a user
      description: Logs out the current user, invalidating their session token
      tags:
        - User
      parameters:
        - name: token
          in: header
          description: user’s current session token
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
  /v1/user/statistics:
    get:
      summary: Returns a user's order statistics
      description: Retrieves statistics to display on the user's dashboard
      tags:
        - User
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatistics'
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
...
  /v1/user/details:
    put:
      summary: Updates a user's details
      description: Updates the details of the current user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: "John"
                lastName:
                  type: string
                  example: "Smith"
                password:
                  type: string
                  example: "Password123"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"

    get: 
      summary: Gets a users details
      description: Get the details of the current user
      tags:
        - User
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  firstName:
                    type: string
                    example: "John"
                  lastName:
                    type: string
                    example: "Smith"
                  password:
                    type: string
                    example: "Password123"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
  /v1/user/reset:
    post:
      summary: Reset a user's password
      description: Allows a user to reset their password
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: integer
                  example: 1234
                newPassword:
                  type: string
                  example: "Password123"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
  /v1/user/forgot:
    post:
      summary: Sends code for password reset
      description: Allows a user to reset their password
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
components:
  responses:
    BadRequestResponse:
      description: Client error, bad request
      content:
        application/json:
          schema:
            type: object
            example:
              error: "Client error, bad request"
    UnauthorizedResponse:
      description: Unauthorized request
      content:
        application/json:
          schema:
            type: object
            example:
              error: "Unauthorized request"
  schemas:
    OrderStatistics:
      type: object
      properties:
        topThreeItems:
          type: array
          items:
            type: string
        numOrdersMonthly:
          type: array
          items:
            type: integer
        totalOrders:
          type: integer
        totalAmountMonth:
          type: number
          format: float
      example:
        topThreeItems: ["item1", "item2", "item3"]
        numOrdersMonthly: [10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65]
        totalOrders: 1000
        totalAmountMonth: 2500.75
...